fixed somehow

+++ Includes/structs.h
@@ -6,7 +6,7 @@
 /*   By: lmicheli <lmicheli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/13 18:07:39 by lmicheli          #+#    #+#             */
-/*   Updated: 2024/05/21 12:50:05 by lmicheli         ###   ########.fr       */
+/*   Updated: 2024/05/21 18:14:07 by lmicheli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,8 +47,8 @@ typedef struct s_pos
 
 typedef struct s_player
 {
-	int			x;
-	int			y;
+	double		x;
+	double		y;
 	double		angle;
 	t_pos		pos; // 0, 50, 0
 	t_pos		dir; // 0, 50, 2
